Put Logic in own package
